🚚 Rename: 아바타 파일 위치 변경 (#107)
