+++ src/components/index.tsx
@@ -15,7 +15,7 @@ import Chat from '@/components/chat/Chat'
 import ListItemSkeleton from '@/components/common/Skeleton/ListItemSkeleton'
 import ImageCardSkeleton from '@/components/common/Skeleton/ImageCardSkeleton'
 import AvatarSkeleton from '@/components/common/Skeleton/AvatarSkeleton'
-import Avatar from '@/components/avatar/Avatar'
+import Avatar from '@/components/common/Avatar'
 import AuthLayout from '@/components/layout/AuthLayout'
 import Notification from '@/components/notification/Notification'
 import { UserInfoDescription } from '@/components/mypage/InfoDescription'
